+++ tests/unit/test_exceptions.py
@@ -1,3 +1,5 @@
+import pickle
+
 from isort import exceptions
 
 
@@ -7,6 +9,9 @@ class TestISortError:
 
     def test_init(self):
         assert isinstance(self.instance, exceptions.ISortError)
+
+    def test_pickleable(self):
+        assert isinstance(pickle.loads(pickle.dumps(self.instance)), exceptions.ISortError)
 
 
 class TestExistingSyntaxErrors(TestISortError):